change: replace search component with search form
