+++ lib/components/Search/Search.stories.tsx
@@ -1,11 +1,14 @@
 import React from 'react'
-import type { Meta, StoryObj } from '@storybook/react'
-import { Search } from './Search'
 import { useCallback, useState } from 'react'
+import { Meta, StoryObj } from '@storybook/react'
+import { Search } from './Search'
 import { SearchDatabase } from '../../data/SearchData'
+import { Modal } from '../Modal/Modal'
+import { useEffect } from 'react'
+import { Button } from '../Button/Button'
 
 const meta: Meta<typeof Search> = {
-  title: 'Components/Search',
+  title: 'Components/Search Form',
   component: Search,
   tags: ['autodocs'],
   parameters: {
@@ -19,42 +22,86 @@ export default meta
 
 type Story = StoryObj<typeof Search>
 
-export const SearchDefault: Story = () => {
-  const [, setOpen] = useState(false)
+export const Default: Story = () => {
+  const [, setMessage] = useState('')
 
   const callback = useCallback(
-    (itemOpen: boolean) => {
-      setOpen(itemOpen)
+    (message: string) => {
+      setMessage(message)
     },
-    [setOpen],
+    [setMessage],
   )
-  return <Search sourceData={SearchDatabase} callback={callback} />
+  return <Search callback={callback} placeholder="Enter a search string" />
 }
 
-export const ValidationUrl: Story = () => {
-  const [, setOpen] = useState(false)
+export const StoryList: Story = () => {
+  const [, setMessage] = useState('')
+  const [searchString, setSearchString] = useState('')
+  const [filteredResults, setFilteredResults] = useState()
+  const searchOn = 'title'
 
   const callback = useCallback(
-    (itemOpen: boolean) => {
-      setOpen(itemOpen)
+    (message: string) => {
+      setMessage(message)
+      setSearchString(message)
     },
-    [setOpen],
+    [setMessage],
   )
-  return <Search sourceData={SearchDatabase} searchOn="url" callback={callback} />
+
+  useEffect(() => {
+    const filteredDatabase =
+      searchString === ''
+        ? []
+        : SearchDatabase.filter((data) => {
+            return data[searchOn].toString().toLowerCase().includes(searchString.toLowerCase())
+          })
+    setFilteredResults(filteredDatabase)
+  }, [searchString])
+
+  return (
+    <Search callback={callback} placeholder="Enter a search string">
+      <Search.Results resultsData={filteredResults} />
+    </Search>
+  )
 }
 
-export const ValidationKey: Story = () => {
-  const [, setOpen] = useState(false)
+export const StoryModal: Story = () => {
+  const [, setMessage] = useState('')
+  const [searchString, setSearchString] = useState('')
+  const [filteredResults, setFilteredResults] = useState()
+  const [modalOpen, setModalOpen] = useState(false)
+  const searchOn = 'title'
 
   const callback = useCallback(
-    (itemOpen: boolean) => {
-      setOpen(itemOpen)
+    (message: string) => {
+      setMessage(message)
+      setSearchString(message)
     },
-    [setOpen],
+    [setMessage],
   )
-  return <Search sourceData={SearchDatabase} searchOn="id" callback={callback} />
+
+  useEffect(() => {
+    const filteredDatabase =
+      searchString === ''
+        ? []
+        : SearchDatabase.filter((data) => {
+            return data[searchOn].toString().toLowerCase().includes(searchString.toLowerCase())
+          })
+    setFilteredResults(filteredDatabase)
+  }, [searchString])
+
+  return (
+    <>
+      <Button title="Click to Open Search" onClick={() => setModalOpen(true)} />
+      <Modal isOpen={modalOpen} setIsOpen={setModalOpen} alignTop>
+        <Search callback={callback} placeholder="Enter a search string">
+          <Search.Results resultsData={filteredResults} />
+        </Search>
+      </Modal>
+    </>
+  )
 }
 
-SearchDefault.storyName = 'Default Search'
-ValidationUrl.storyName = 'Validation Url Search'
-ValidationKey.storyName = 'Validation Key Search'
+Default.storyName = 'Default Search Form'
+StoryList.storyName = 'Search Form with List'
+StoryModal.storyName = 'Search Form with List in Modal'
